Move class FitEllipsoidPluginDemo to new sub-package demo
